♻️ split stacked dispatch responsibilities
